Only set GS_APP_SPECIAL_KIND_OS_UPGRADE on non-appstream apps
The purpose of that flag is to mark updates of components which
do not have an associated AppStream entry. That is done by setting
the component kind to AS_COMPONENT_KIND_GENERIC and setting a value
in GsAppSpecialKind. Doing it differently creates a bug where appstream
apps are not correctly identified as having to be upgraded or downgraded.