gs-plugin-apk: fix leak in g_variant_to_apkd_package
The g_variant_get_child_value calls create a new GVariant which
is expected to be freed[1]. Right now, by piping the result, such
GVariant is directly leaked. Instead, use g_variant_get with the
signature of the variant. Use "&s" for every string, to avoid
generating a copy and getting ownership[2]. This way, all the data
actually resides in the value_tuple parameter.

[1]https://docs.gtk.org/glib/method.Variant.get_child_value.html#return-value
[2]https://docs.gtk.org/glib/gvariant-format-strings.html#pointers