gs-plugin-apk: steal local pointers returned in async functions
Fixes #49

The local errors are defined like g_autoptr (GError). Therefore,
they will be freed at the end of the function. Since the error
might be used in the finish part of the async function and has to
be propagated. Freeing the errors causes an use-after-free bug
that later sometimes leads to Segmentation Faults deep into the
glib stack.