gs-plugin-apk: remove reporting of app progress
The current implementation has not worked too well for a while,
and is totally problematic since the plugin was ported to 42, where
multiple calls could be happening at the same time and races happen
all the time. Since this is simply a cosmetic feature, remove it for
now until there is the interest, time and knowledge to maintain it.