Aggregate all calls to SearchFilesOwners and add tests
When there are multiple apps which do not have a package assigned,
the initial refine and the upgrades can take unnecessary time.
Solve this problem by making use of the ApkPolkit2 functionality that
allows multiple searches in a single call. Also, add tests to
help verify correctness of the changes.